feat: update .gitignore and add dataset upload script for WandB

+++ e3_diffusion_for_molecules/qm9/data/prepare/qm9.py
@@ -28,12 +28,11 @@ def download_dataset_qm9(datadir, dataname, splits=None, calculate_thermo=True, 
     logging.info('Beginning download of GDB9 dataset!')
     gdb9_url_data = 'https://springernature.figshare.com/ndownloader/files/3195389'
     gdb9_tar_data = join(gdb9dir, 'dsgdb9nsd.xyz.tar.bz2')
-    # gdb9_tar_file = join(gdb9dir, 'dsgdb9nsd.xyz.tar.bz2')
-    # gdb9_tar_data =
-    # tardata = tarfile.open(gdb9_tar_file, 'r')
-    # files = tardata.getmembers()
-    urllib.request.urlretrieve(gdb9_url_data, filename=gdb9_tar_data)
-    logging.info('GDB9 dataset downloaded successfully!')
+    if not os.path.exists(gdb9_tar_data):
+        urllib.request.urlretrieve(gdb9_url_data, filename=gdb9_tar_data)
+        logging.info('GDB9 dataset downloaded successfully!')
+    else:
+        logging.info('GDB9 dataset already exists, skipping download.')
 
     # If splits are not specified, automatically generate them.
     if splits is None:
@@ -82,7 +81,8 @@ def gen_splits_gdb9(gdb9dir, cleanup=True):
     logging.info('Splits were not specified! Automatically generating.')
     gdb9_url_excluded = 'https://springernature.figshare.com/ndownloader/files/3195404'
     gdb9_txt_excluded = join(gdb9dir, 'uncharacterized.txt')
-    urllib.request.urlretrieve(gdb9_url_excluded, filename=gdb9_txt_excluded)
+    if not os.path.exists(gdb9_txt_excluded):
+        urllib.request.urlretrieve(gdb9_url_excluded, filename=gdb9_txt_excluded)
 
     # First get list of excluded indices
     excluded_strings = []
@@ -147,7 +147,8 @@ def get_thermo_dict(gdb9dir, cleanup=True):
     gdb9_url_thermo = 'https://springernature.figshare.com/ndownloader/files/3195395'
     gdb9_txt_thermo = join(gdb9dir, 'atomref.txt')
 
-    urllib.request.urlretrieve(gdb9_url_thermo, filename=gdb9_txt_thermo)
+    if not os.path.exists(gdb9_txt_thermo):
+        urllib.request.urlretrieve(gdb9_url_thermo, filename=gdb9_txt_thermo)
 
     # Loop over file of thermochemical energies
     therm_targets = ['zpve', 'U0', 'U', 'H', 'G', 'Cv']
